+++ src/modules/commentPopup.js
@@ -1,11 +1,27 @@
-/* eslint-disable  import/no-cycle */
+const commentCounter = (data) => (typeof (data) === 'object' ? data.length : 'invalid');
 
-import getMovieComment from './comment.js';
-import { updateCommentCounter } from './commentCounter.js';
+const commentsApiKey = 'https://us-central1-involvement-api.cloudfunctions.net/capstoneApi/apps/dLRWdvDoWjaapH1JgaCf/comments';
+const getMovieComment = async (movieId) => {
+  const response = await fetch(`${commentsApiKey}?item_id=${movieId}`);
+  return response.json();
+};
+
+const commentPopup = document.querySelector('.comment-popup');
+const getTotalComments = async (movieId) => {
+  const result = await getMovieComment(movieId)
+    .then((comment) => (!comment.error ? comment.length : 0))
+    .catch(() => 0);
+  return result;
+};
+
+const updateCommentCounter = (movieId) => {
+  getTotalComments(movieId).then((totalComment) => {
+    commentPopup.querySelector('.total-comments').innerHTML = totalComment;
+  });
+};
 
 const commentApiEndpoint = 'https://us-central1-involvement-api.cloudfunctions.net/capstoneApi/apps/dLRWdvDoWjaapH1JgaCf/comments';
 const movieApi = 'https://api.tvmaze.com/shows';
-const commentPopup = document.querySelector('.comment-popup');
 
 const get = (url) => fetch(url)
   .then((res) => res.json())
@@ -123,4 +139,13 @@ const showCommentPopup = async (movieId) => {
   closeCommentPopup();
 };
 
-export default showCommentPopup;
+document.addEventListener('click', async (e) => {
+  if (e.target.matches('.comment-btn')) {
+    showCommentPopup(e.target.id);
+    const comment = await getMovieComment(e.target.id);
+    commentCounter(comment);
+    updateCommentCounter(e.target.id);
+  }
+});
+
+export { showCommentPopup, commentCounter };
